requirements.txt
alembic==1.7.4
aniso8601==8.0.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
distlib==0.3.3
filelock==3.3.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.21.3
pipenv==2021.5.29
platformdirs==2.4.0
psycopg2==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
SQLAlchemy==1.3.18
twilio==7.2.0
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.2
